requirements.txt
mss
openai
python-dotenv
google-genai
pyqt5
keyboard
RealtimeSTT
google-cloud-speech
sounddevice
scipy